vss-git-convert

+++ SHARED/modAPChk.bas
@@ -5,17 +5,26 @@ Option Explicit
 '#Ticket #: 417974. WJ 10/27/2003
 '#In what follows, we implement a routine to lock/unlock p_checks.
 '#It also generate check nbr or valid check #
-'#The stored precedure get_chcke_nbr did not work correctly.
+'#The stored precedure get_chcke_nbr did not work correctly since it did
+'#not ask check account. This module will be used in all VB AP check #
+'#validations.
 '#
+'# How to use this module
+'#(1)Call fnLockP_Checks(Prt_grp,User,lstart,bSingleCheck) in check validation
+'#   if only one check will be printed , pass bSingleCheck= true
+'#   if want to generate start check, user lStart = 0 , a valid start check
+'#   will be returned(byref). If check is locked sucessfully, fnLockP_Checks=SZEMPTY
+'#   otherwise, error message will be returned.
+'#(2)call subUnlockP_checks after AP check is done or Canceled
 '###################################################################
 Private m_FirstLockedCheck As String
 Private m_LastLockedCheck As String
 
-Private Function fnFormatPCheck(ByVal lAcct As Long, ByVal lChk As Long)
+Private Function fnFormatPCheck(ByVal lAcct As Long, ByVal lChk As Long) As String
     fnFormatPCheck = Format(lAcct, String(10, "0")) & Format(lChk, String(10, "0"))
 End Function
 
-Private Function fnLockP_Checks(ByVal sPGrp As String, _
+Public Function fnLockP_Checks(ByVal sPGrp As String, _
                                 ByVal sUser As String, _
                                 ByRef lStart As Long, _
                                 Optional bSingleCheck As Boolean = False) As String
@@ -30,10 +39,16 @@ Private Function fnLockP_Checks(ByVal sPGrp As String, _
     Dim rsTemp As Recordset
     Dim lNumberChecks As Long
     Dim lRecCount As Long
+    Dim bGenerateStartChk As Boolean
+    Dim sLockData As String * 20
+    Dim lReservedNbr As Long
+    
+    On Error GoTo errTrap
     
     '#Make sure the Print Group is Valid,
     '#then get acct from p group
     sPGrp = Trim(sPGrp)
+    sUser = Trim(sUser)
     
     strSQL = "SELECT pg_chk_acct FROM p_print WHERE pg_group = " & tfnSQLString(sPGrp)
     
@@ -74,9 +89,18 @@ Private Function fnLockP_Checks(ByVal sPGrp As String, _
             Exit Function
         End If
      End If
+      
      
+     '#Unlock any checks done preciously
+     subUnlockP_checks
+    
      '#If we need to generate check, then do it
      If lStart = 0 Then
+        bGenerateStartChk = True
+     Else
+        bGenerateStartChk = False
+     End If
+     If bGenerateStartChk Then
         strSQL = "SELECT max(pv_check_nbr) max_check_nbr from p_checks WHERE pv_account = " & lChkAcct
         lRecCount = apc_GetRecordSet(rsTemp, strSQL)
         If lRecCount < 0 Then
@@ -84,58 +108,57 @@ Private Function fnLockP_Checks(ByVal sPGrp As String, _
              Exit Function
         End If
         lStart = tfnRound(rsTemp!max_check_nbr) + 1
-     End If
-    
-    'Validate check #s against p_checks!
-    
-    lEnd = lStart + lNumberChecks - 1
-    strSQL = "SELECT pv_check_nbr FROM p_checks WHERE pv_account = " & lChkAcct _
-           & " AND pv_check_nbr BETWEEN " & CStr(lStart) & " AND " & CStr(lEnd)
-    
-    lRecCount = apc_GetRecordSet(rsTemp, strSQL)
-    If lRecCount < 0 Then
-         fnLockP_Checks = QUERY_FAILED
-         Exit Function
-    End If
-    If lRecCount > 0 Then
-        fnLockP_Checks = "Check # " & tfnRound(rsTemp!pv_check_nbr) & " is used; it falls between start and end numbers"
-        Exit Function
-    End If
-    
-    '#Unlock any checks done preciously
-    subUnlockP_checks
-    
-    '#Check the if any of the check is locked
-    sTemp = "'" & fnFormatPCheck(lChkAcct, lStart) & "'"
-    j = 0
-    For i = lStart + 1 To lEnd
-        j = j + 1
-        sTemp = sTemp & ",'" & fnFormatPCheck(lChkAcct, i) & "'"
-        If i = lEnd Or j = 500 Then
-            strSQL = "SELECT srl_criteria FROM sys_row_lock" _
-                   & " WHERE srl_table = '" & TABLENAME & "'" _
-                   & " AND srl_criteria IN (" & sTemp & ")"
-            lRecCount = apc_GetRecordSet(rsTemp, strSQL)
-            If lRecCount < 0 Then
-                 fnLockP_Checks = QUERY_FAILED
-                 Exit Function
-            ElseIf lRecCount > 0 Then
-                fnLockP_Checks = "Check # " & tfnRound(Right(rsTemp!srl_criteria & "", 10)) & " is locked; it falls between start and end numbers"
-                Exit Function
+        
+        '#Check if any chk # are reserved
+        strSQL = "SELECT srl_criteria FROM sys_row_lock" _
+                & " WHERE srl_table = '" & TABLENAME & "'" _
+                & " AND srl_criteria[1,10] = '" & Format(lChkAcct, String(10, "0")) & "'"
+        strSQL = strSQL & " ORDER BY srl_criteria DESC"
+        If apc_GetRecordSet(rsTemp, strSQL) > 0 Then
+            sLockData = rsTemp!srl_criteria & ""
+            lReservedNbr = Val(Mid(sLockData, 11, 10))
+            If lStart <= lReservedNbr Then
+                lStart = lReservedNbr + 1
             End If
-            j = 0
-            sTemp = "'" & fnFormatPCheck(lChkAcct, lStart) & "'"
         End If
-    Next i
+        lEnd = lStart + lNumberChecks - 1
+     Else
+        '#Validate check #s against p_checks!
+        lEnd = lStart + lNumberChecks - 1
+        strSQL = "SELECT pv_check_nbr FROM p_checks WHERE pv_account = " & lChkAcct _
+               & " AND pv_check_nbr BETWEEN " & CStr(lStart) & " AND " & CStr(lEnd)
+        
+        lRecCount = apc_GetRecordSet(rsTemp, strSQL)
+        If lRecCount < 0 Then
+             fnLockP_Checks = QUERY_FAILED
+             Exit Function
+        End If
+        If lRecCount > 0 Then
+            fnLockP_Checks = "Check # " & tfnRound(rsTemp!pv_check_nbr) & " is used; it falls between start and end numbers"
+            Exit Function
+        End If
     
+        '#Check the if any of the check is locked
+        strSQL = "SELECT srl_criteria FROM sys_row_lock" _
+                & " WHERE srl_table = '" & TABLENAME & "'" _
+                & " AND srl_criteria[1,10] = '" & Format(lChkAcct, String(10, "0")) & "'"
+        strSQL = strSQL & " AND srl_criteria BETWEEN '" & fnFormatPCheck(lChkAcct, lStart) & "'"
+        strSQL = strSQL & " AND '" & fnFormatPCheck(lChkAcct, lEnd) & "'"
+        If apc_GetRecordSet(rsTemp, strSQL) > 0 Then
+            fnLockP_Checks = "Check # " & tfnRound(Right(Trim(rsTemp!srl_criteria & ""), 10)) & " is locked; it falls between start and end numbers"
+            Exit Function
+        End If
+    End If
     '#Lock the Checks
     strSQL = ""
     j = 0
     For i = lStart To lEnd
         j = j + 1
-        strSQL = strSQL & "INSERT INTO sys_row_lock VALUES('" & TABLENAME & "'," & tfnSQLString(LCase(App.EXEName)) & ",'" & tfnGetUserName & "'," & fnFormatPCheck(lChkAcct, i) & ",0);"
+        strSQL = strSQL & "INSERT INTO sys_row_lock VALUES('" & TABLENAME & "'," & tfnSQLString(LCase(App.EXEName)) & ",'" & tfnGetUserName & "','" & fnFormatPCheck(lChkAcct, i) & "',0);"
         If i = lEnd Or j = 300 Then
             If Not apc_ExecuteSQL(strSQL) Then
+                m_FirstLockedCheck = fnFormatPCheck(lChkAcct, lStart)
+                m_LastLockedCheck = fnFormatPCheck(lChkAcct, lEnd)
                 fnLockP_Checks = QUERY_FAILED
                 Exit Function
             End If
@@ -146,13 +169,20 @@ Private Function fnLockP_Checks(ByVal sPGrp As String, _
     
     m_FirstLockedCheck = fnFormatPCheck(lChkAcct, lStart)
     m_LastLockedCheck = fnFormatPCheck(lChkAcct, lEnd)
+    
+    fnLockP_Checks = szEMPTY
+    Exit Function
+errTrap:
+    fnLockP_Checks = "Error: " & Err.Description
 End Function
 
-Private Sub subUnlockP_checks()
+Public Sub subUnlockP_checks()
     Const SUB_NAME = "subUnlockP_checks"
     Dim sSql As String
     Dim rsTemp As Recordset
-    
+    If m_FirstLockedCheck = "" Then
+       Exit Sub
+    End If
     sSql = "DELETE FROM sys_row_lock" _
          & " WHERE srl_table = 'p_checks'" _
          & " AND srl_prog_id = " & tfnSQLString(LCase(App.EXEName)) _
